models/requirements.txt
# Python 3.8.11
numpy~=1.16.6
PyYAML~=5.4.1
pydantic==1.8.2
tqdm~=4.56.2
scikit-learn~=0.24.2
spacy==3.2.0    # 3.0.6
rasa==2.3.5
MarkupSafe==2.0.1
torch==1.6.0
transformers[sentencepiece]==4.4.2